tests: Switch non-parallel learner tests to scikit-learn from spams.
The point of this change is to begin the transition to scikit-learn for
dictionary learning. This is a partial transition as scikit-learn does
not behave correctly in the parallel cases due to an incorrect sign.
Adding a positivity constraint to dictionary learning would correct this
problem, but this has not yet been accepted. So, SPAMS will continue to
be used for those the parallel tests.