tests: Skip tests requiring SPAMS to be installed.
Some dictionary learning test for SPAMS are required to make sure that
code behaves correctly. However, there is no point in running them when
SPAMS is not installed. Also, the parallel learner regression tests fail
to work correctly with scikit-learn at present. These continue to run
with SPAMS for the interim. However, if SPAMS is not present for some
test cases (e.g. Python 3), these should simply be skipped. Once a fix
for scikit-learn is made so it will work in these cases it will be
switched to instead.