tests: Switch from `multiprocessing` to `queue`/`Queue` for testing.
Appears that this will hang sometimes if we use a multiprocessing Queue
in the same process. It is unclear exactly what causes the problem to
arise or not. Instead of trying to determine the cause simply switch to
a normal `Queue`, which should not deadlock in the same process and will
be just as effective for these tests.