Remove nanshe-viewer as it is no longer used.
We plan on moving the history of this addition to an archive repo where
it will not be actively developed. In any event, it is still safe for
us to eliminate it from this repo as the history still remains
available. README changes will be handled in a subsequent PR.