Add sorted axes in `unsquish` to an `OrderedDict`.
It seems how `dict`s are created an sorted has changed in Python 3.6.
This causes a bug in `unsquish`, which relied on the keys being sorted.
To fix this, we go ahead and sort the axes in the intended way first.
Then we add them to an `OrderedDict` to preserve the ordering. By using
this strategy, we do this in a way we ensure are constraints are met
without relying on an implementation detail of `dict`.