Move the coverage erase setting to setup.cfg
Newer versions of `coverage` don't accept the `erase` option in `run` of
the `.coveragerc` configuration file. So just use `nose`'s option for
this in `setup.cfg` and `setup.all.cfg`. This will have the same effect
and will make it easier for us to support `coverage` 4.* instead of
being stuck on 3.* as we are.