Drop HDF5 pin
This was more important before channel priority was a thing in `conda`
to make sure that we didn't get a copy of HDF5 that was ABI
compatibility with our other libraries. Now that channel priority works
and basically everything we use is in conda-forge using even newer
versions of HDF5, we should drop this pinning. After all this is a pure
Python library that only interfaces with HDF5 through other Python
libraries (e.g. h5py), so pinning doesn't make much sense for any other
reason here. Not to mention this pinning due to other related
constraints was keeping us on NumPy 1.12 instead of NumPy 1.13 (latest),
which we now support.