Use NumPy's legacy 1.13 printing
Our doctests are formatted for NumPy 1.13. However that means some of
the fail on NumPy 1.14. Fortunately NumPy 1.14 added a legacy print
option that allows us to force the NumPy 1.13 style of printing. Hence
we enable this for module's whose doctests fail on NumPy 1.14.
Unfortunately this legacy option is new in NumPy 1.14 (i.e. does not
exist in NumPy 1.13), so catch the `TypeError` exception one would see
on NumPy 1.13 or earlier.