Drop the Matplotlib-based viewer
This has since been refactored out as the package `mplview`, which we
have cleaned up and added testing for. The implementation here is not
actually tested or really used as far as I know. Not to mention it has
been deprecated for a long time. We have also long since aliased
`mplview`'s classes to match the old class names here. So users have had
a while to migrate to `mplview` and those that are lagging don't have to
do too much work to fix it. Given this go ahead and drop our viewer.