Require NumPy 1.11+
This is a workaround for a conda-build bug in `bdist_conda`. However
having a constraint for NumPy here is reasonable. Though the code was
designed for such an old version of NumPy that most versions likely
work without issues. For now we make the requirement 1.11 as we are
unlikely to go further back than that.